Moving digest algorithm and factory to crypto extension
