+++ cordapp-configuration/src/main/resources/net/corda/cordapp/cordapp-configuration.properties
@@ -6,7 +6,7 @@ Corda-CheckpointCustomSerializer-Classes=IMPLEMENTS;net.corda.v5.serialization.C
 Corda-NotaryService-Classes=EXTENDS;net.corda.v5.ledger.notary.NotaryService
 Corda-StateAndRefPostProcessor-Classes=IMPLEMENTS;net.corda.v5.ledger.services.vault.StateAndRefPostProcessor
 Corda-CustomQueryPostProcessor-Classes=IMPLEMENTS;net.corda.v5.application.persistence.query.CustomQueryPostProcessor
-Corda-DigestAlgorithmFactory-Classes=IMPLEMENTS;net.corda.v5.cipher.suite.DigestAlgorithmFactory
+Corda-DigestAlgorithmFactory-Classes=IMPLEMENTS;net.corda.v5.crypto.extensions.DigestAlgorithmFactory
 Corda-Entity-Classes=ANNOTATED;javax.persistence.Entity;ANNOTATED;net.corda.v5.base.annotations.CordaSerializable
 Corda-RPCStartableFlow-Classes=IMPLEMENTS;net.corda.v5.application.flows.RPCStartableFlow
 Corda-InitiatedFlow-Classes=IMPLEMENTS;net.corda.v5.application.flows.ResponderFlow
