+++ crypto-extensions/src/main/java/net/corda/v5/crypto/extensions/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.v5.crypto.extensions;
+
+import org.osgi.annotation.bundle.Export;
