+++ settings.gradle
@@ -92,6 +92,7 @@ include 'cipher-suite'
 include 'corda-api'
 include 'cordapp-configuration-publish'
 include 'crypto'
+include 'crypto-extensions'
 include 'data:avro-schema'
 include 'data:config-schema'
 include 'data:db-schema'
